feat: all of the vehicle endpoints implement the jwt validation and display different behaviours between the private user/corporate user/corporate admin/root

+++ rest-api/internal/handlers/vehicle/vehicle.go
@@ -10,11 +10,29 @@ import (
 	"github.com/czxrny/veh-sense-backend/shared/models"
 )
 
+// Returns all the vehicles that are available to user:
+// If the user is a private user: returns all of the private cars
+// If the user is corporate and is not an admin: returns all of the assigned cars + shared (the ones without the owner_id)
+// If the user is an admin of organization: returns all of the organization cars
+// Root gets all the info.
 func GetVehicles(w http.ResponseWriter, r *http.Request) {
-	/* TODO - RETURN ONLY THE VEHICLES FROM THE ORGRANIZATION/PRIVATE OWNER */
-	/* POSSIBLE USAGE - OWNER WILL HAVE MULTIPLE VEHICLES THAT CAN BE DISPLAYED UPON THE START OF THE APP */
 	common.GetAllHandler(w, r, func(ctx context.Context) ([]models.Vehicle, error) {
+		authClaims, ok := ctx.Value("authClaims").(models.AuthInfo)
+		if !ok {
+			return nil, fmt.Errorf("Error: Internal server error. Something went wrong while decoding the JWT.")
+		}
+
 		db := database.GetDatabaseClient()
+		switch authClaims.Role {
+		case "user":
+			db = db.Where("owner_id = ?", authClaims.UserID)
+			// For shared vehicles - if the user is corporate
+			if authClaims.OrganizationID != nil {
+				db = db.Or("organization_id = ? AND owner_id IS NULL", authClaims.OrganizationID)
+			}
+		case "admin":
+			db = db.Where("organization_id = ?", authClaims.UserID)
+		}
 
 		var vehicles []models.Vehicle
 		if err := db.Find(&vehicles).Error; err != nil {
@@ -27,6 +45,29 @@ func GetVehicles(w http.ResponseWriter, r *http.Request) {
 
 func AddVehicle(w http.ResponseWriter, r *http.Request) {
 	common.PostHandler(w, r, func(ctx context.Context, vehicle *models.Vehicle) (*models.Vehicle, error) {
+		authClaims, ok := ctx.Value("authClaims").(models.AuthInfo)
+		if !ok {
+			return nil, fmt.Errorf("Error: Internal server error. Something went wrong while decoding the JWT.")
+		}
+
+		switch authClaims.Role {
+		case "user":
+			if authClaims.OrganizationID != nil {
+				return nil, fmt.Errorf("Error: Corporate user is unauthorized to add new vehicles to fleet. Login as an admin to proceed.")
+			}
+			vehicle.OwnerID = &authClaims.UserID
+			vehicle.OrganizationID = authClaims.OrganizationID
+
+		// Admin sets the vehicle organization_id automatically, and can pass the owner_id to specify the user that will be the owner
+		case "admin":
+			vehicle.OrganizationID = authClaims.OrganizationID
+
+		case "root":
+			if vehicle.OwnerID == nil || vehicle.OrganizationID == nil {
+				return nil, fmt.Errorf("Error: Bad Request: Please specify either the organization_id or owner_id to proceed")
+			}
+		}
+
 		db := database.GetDatabaseClient()
 		if err := db.Create(vehicle).Error; err != nil {
 			return nil, err
@@ -36,9 +77,14 @@ func AddVehicle(w http.ResponseWriter, r *http.Request) {
 	})
 }
 
+// Returns the vehicle - only if the user is the owner of the vehicle / the vehicle is shared in the corporation the user is in
 func GetVehicleById(w http.ResponseWriter, r *http.Request) {
-	/* PROVIDE ONLY IF THE USER IS THE OWNER! */
 	common.GetByIdHandler(w, r, func(ctx context.Context, id int) (*models.Vehicle, error) {
+		authClaims, ok := ctx.Value("authClaims").(models.AuthInfo)
+		if !ok {
+			return nil, fmt.Errorf("Error: Internal server error. Something went wrong while decoding the JWT.")
+		}
+
 		db := database.GetDatabaseClient()
 
 		var vehicle models.Vehicle
@@ -46,24 +92,43 @@ func GetVehicleById(w http.ResponseWriter, r *http.Request) {
 			return nil, err
 		}
 
+		isOwner := vehicle.OwnerID != nil && *vehicle.OwnerID == authClaims.UserID
+		isShared := vehicle.OrganizationID != nil && authClaims.OrganizationID != nil && *vehicle.OrganizationID == *authClaims.OrganizationID && vehicle.OwnerID == nil
+		isOrgAdmin := vehicle.OrganizationID != nil && authClaims.OrganizationID != nil && *vehicle.OrganizationID == *authClaims.OrganizationID && authClaims.Role == "admin"
+
+		if !isOwner && !isShared && !isOrgAdmin && authClaims.Role != "root" {
+			return nil, fmt.Errorf("Error: User is unauthorized to view the vehicle.")
+		}
+
 		return &vehicle, nil
 	})
 }
 
 func UpdateVehicle(w http.ResponseWriter, r *http.Request) {
-	/* TODO - ONLY THE ORGANIZATION ADMIN / OWNER CAN EDIT THE VEHICLE INFO.. */
 	common.PatchHandler(w, r, func(ctx context.Context, updatedVehicle *models.VehicleUpdate, id int) (*models.Vehicle, error) {
+		authClaims, ok := ctx.Value("authClaims").(models.AuthInfo)
+		if !ok {
+			return nil, fmt.Errorf("Error: Internal server error. Something went wrong while decoding the JWT.")
+		}
+
 		db := database.GetDatabaseClient()
+		var vehicle models.Vehicle
+		if err := db.First(&vehicle, id).Error; err != nil {
+			return nil, err
+		}
+
+		isOwner := vehicle.OwnerID != nil && *vehicle.OwnerID == authClaims.UserID
+		isOrgAdmin := vehicle.OrganizationID != nil && authClaims.OrganizationID != nil && *vehicle.OrganizationID == *authClaims.OrganizationID && authClaims.Role == "admin"
+
+		if !isOwner && !isOrgAdmin && authClaims.Role != "root" {
+			return nil, fmt.Errorf("Error: User is unauthorized to edit the vehicle.")
+		}
+
 		result := db.Model(&models.Vehicle{}).Where("id=?", id).Updates(updatedVehicle)
 		if result.Error != nil {
 			return nil, result.Error
 		}
 
-		if result.RowsAffected == 0 {
-			return nil, fmt.Errorf("Vehicle does not exist!")
-		}
-
-		var vehicle models.Vehicle
 		if err := db.First(&vehicle, id).Error; err != nil {
 			return nil, err
 		}
@@ -74,14 +139,27 @@ func UpdateVehicle(w http.ResponseWriter, r *http.Request) {
 
 func DeleteVehicle(w http.ResponseWriter, r *http.Request) {
 	common.DeleteHandler(w, r, func(ctx context.Context, id int) error {
+		authClaims, ok := ctx.Value("authClaims").(models.AuthInfo)
+		if !ok {
+			return fmt.Errorf("Error: Internal server error. Something went wrong while decoding the JWT.")
+		}
+
 		db := database.GetDatabaseClient()
+		var vehicle models.Vehicle
+		if err := db.First(&vehicle, id).Error; err != nil {
+			return err
+		}
+
+		isOwner := vehicle.OwnerID != nil && *vehicle.OwnerID == authClaims.UserID
+		isOrgAdmin := vehicle.OrganizationID != nil && authClaims.OrganizationID != nil && *vehicle.OrganizationID == *authClaims.OrganizationID && authClaims.Role == "admin"
+
+		if !isOwner && !isOrgAdmin && authClaims.Role != "root" {
+			return fmt.Errorf("Error: User is unauthorized to delete the vehicle.")
+		}
+
 		result := db.Delete(&models.Vehicle{}, id)
 		if result.Error != nil {
 			return result.Error
-		}
-
-		if result.RowsAffected == 0 {
-			return fmt.Errorf("No vehicle found to delete")
 		}
 
 		return nil
